hello/Youmoney-task/requirements.txt
yookassa
python-dotenv
flask
Flask-RESTful
marshmallow
celery
redis
flower
python-telegram-bot